refactor: src/page 폴더내에서 도메인 로직을 제거하기 위해, RoadmapDetail 컴포넌트 생성
